parseSitemap advances to next rule.
parseSitemap was a recent addition. I chose not to thoroughly test
it because it wasn't a part of any of the core examples in Google's
specification.

However, there is a sitemap rule in the pathological input file.
Upon testing the tool against that, it was clear that the parser
was stuck in a loop. When it parsed a sitemap rule, it didn't advance
to the next rule: it just kept parsing the same sitemap rule
indefinitely.

Lesson: always test with the pathological input!