Lexer supports spaces within agent names.
The grammar for the language specified by Google allowed for
spaces within the name of agents. I'd chosen to ignore this
because I assumed it was complex (given the sort-of-subtlety of
the LWS code). Then it came to me that I could just trim whitespace
from the end of an item's value in the event that there wasn't a comment
on the same line. Done.