r.Test() discards scheme and domain.
Previously r.Test would only accept a path. However, this is a pain
for clients of the library. Instead, an absolute URL can now be
provided. The scheme and domain of the URL, if provided, will be
discarded. Therefore the caller must check the applicability of the
robots.txt file to the provided URL.