Lexer checks for BOM.
I hadn't had any tests for BOM. The pathological test file had a BOM,
but it wasn't immediately followed by a valid rule. So I changed the
file to include such a rule, added a test that would check for success,
and the test failed. So I added a quick function to remove the BOM
before the lexer starts working.