Robots accepts status field.
This is the start of work to have the Robots type respect the status
code behavior in Google's specification. I've started by making a new
type robotsdata. robotsdata is what Robots used to be. Now it is no
longer exported.

Robots now contains robotsdata and a status code. All tests pass;
otherwise behavior is as it was. Next step is to change the test
behavior to reflect the status code.

To keep from exporting the Sitemaps field on robotsdata, I've also added
a Sitemaps method to Robots. This is a better design choice. I should've
started here anyway.