Improved extraction of path from rawurl being tested.
The Robots Test* methods are intended to be easy to use and accept a
wide range of inputs. That means either absolute or relative URLs. Only
the path part of the URL was being matched against.

However, the robots.txt directives can also match against the query part
of a URL. Therefore if the provided URL has a query string, it needs to
be appended to the path being tested. This commit adds an unexported
function that accomplishes that invisibly.