Remove "sort" dependency.
Previously to ensure a match would always be the longest, lists
of agents and members were sorted every time a new "rule" was inserted.
This is obviously unnecessary since we have control over the full
lifecycle of the lists: we can maintain order by inserting new values
wherever they are meant to go. That's what this update does.

Implementing this also happens to remove the dependency of the package
on "sort", which is nice.